requirements.txt
pandas
numpy
pymongo
pymongo[srv]
certifi
python-dotenv

# -e .

# refers to the setup file